Simplify the code (1/x) for #1

+++ load.v
@@ -1,6 +1,6 @@
 module main
 
-fn load_exe(exe_contents string, mut exe_memory []u8, pe32_dos_header PE32_DOS_HEADER, pe32_file_header PE32_FILE_HEADER, pe32_optional_header PE32_OPTIONAL_HEADER, pe32_section_headers []PE32_SECTION_HEADER, pe_header_pointer u16, exe_sections_count int)
+fn load_exe(exe_contents string, mut exe_memory []u8, pe32_dos_header PE32_DOS_HEADER, pe32_file_header PE32_FILE_HEADER, pe32_optional_header PE32_OPTIONAL_HEADER, pe32_section_headers []PE32_SECTION_HEADER, pe_header_pointer u32, exe_sections_count int)
 {
 	// Load headers to EXE exe_memory
